requirements.txt
Django==1.5.1
South==0.8.2
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.31
celery==3.0.21
django-celery==3.0.21
django-model-utils==1.4.0
gunicorn==17.5
kombu==2.5.12
psycopg2==2.5.1
python-dateutil==2.1
pytz==2013b
six==1.3.0
unicodecsv==0.9.4
wsgiref==0.1.2
